Techneers/Generative AI/GenAI/requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
aiostream==0.5.2
altair==4.2.2
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
cassandra-driver==3.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
Faker==23.2.1
fastjsonschema==2.19.1
favicon==0.7.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.41
gradientai==1.7.0
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
langchain==0.0.154
llama-index==0.9.12
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
openai==1.12.0
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7
Pympler==1.0.1
pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.31.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
tiktoken==0.6.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0